+++ eras/shelley/impl/testlib/Test/Cardano/Ledger/Shelley/ImpTest.hs
@@ -147,7 +147,7 @@ module Test.Cardano.Ledger.Shelley.ImpTest (
   withPreFixup,
   impNESL,
   impGlobalsL,
-  impLastTickG,
+  impCurSlotNoG,
   impKeyPairsG,
   impNativeScriptsG,
   produceScript,
@@ -259,7 +259,7 @@ import Data.List.NonEmpty (NonEmpty)
 import Data.Map.Strict (Map)
 import qualified Data.Map.Strict as Map
 import Data.Maybe (catMaybes, isNothing, mapMaybe)
-import Data.Ratio ((%))
+import Data.Ratio (denominator, numerator, (%))
 import Data.Sequence.Strict (StrictSeq (..))
 import qualified Data.Sequence.Strict as SSeq
 import qualified Data.Set as Set
@@ -342,7 +342,7 @@ data ImpTestState era = ImpTestState
   , impKeyPairs :: !(Map (KeyHash Witness) (KeyPair Witness))
   , impByronKeyPairs :: !(Map BootstrapAddress ByronKeyPair)
   , impNativeScripts :: !(Map ScriptHash (NativeScript era))
-  , impLastTick :: !SlotNo
+  , impCurSlotNo :: !SlotNo
   , impGlobals :: !Globals
   , impEvents :: [SomeSTSEvent era]
   }
@@ -385,11 +385,11 @@ impGlobalsL = lens impGlobals (\x y -> x {impGlobals = y})
 impNESL :: Lens' (ImpTestState era) (NewEpochState era)
 impNESL = lens impNES (\x y -> x {impNES = y})
 
-impLastTickL :: Lens' (ImpTestState era) SlotNo
-impLastTickL = lens impLastTick (\x y -> x {impLastTick = y})
+impCurSlotNoL :: Lens' (ImpTestState era) SlotNo
+impCurSlotNoL = lens impCurSlotNo (\x y -> x {impCurSlotNo = y})
 
-impLastTickG :: SimpleGetter (ImpTestState era) SlotNo
-impLastTickG = impLastTickL
+impCurSlotNoG :: SimpleGetter (ImpTestState era) SlotNo
+impCurSlotNoG = impCurSlotNoL
 
 impRootTxInL :: Lens' (ImpTestState era) TxIn
 impRootTxInL = lens impRootTxIn (\x y -> x {impRootTxIn = y})
@@ -639,7 +639,7 @@ defaultInitImpTestState nes = do
       , impKeyPairs = prepState ^. keyPairsL
       , impByronKeyPairs = prepState ^. keyPairsByronL
       , impNativeScripts = mempty
-      , impLastTick = slotNo
+      , impCurSlotNo = slotNo
       , impGlobals = globals
       , impEvents = mempty
       }
@@ -730,7 +730,7 @@ disableInConformanceIt s =
 
 impLedgerEnv :: EraGov era => NewEpochState era -> ImpTestM era (LedgerEnv era)
 impLedgerEnv nes = do
-  slotNo <- gets impLastTick
+  slotNo <- gets impCurSlotNo
   epochNo <- runShelleyBase $ epochFromSlot slotNo
   pure
     LedgerEnv
@@ -1337,11 +1337,22 @@ passTick ::
   ) =>
   ImpTestM era ()
 passTick = do
-  impLastTick <- gets impLastTick
+  impCurSlotNo <- gets impCurSlotNo
   curNES <- getsNES id
-  nes <- runImpRule @"TICK" () curNES impLastTick
-  impLastTickL += 1
+  nes <- runImpRule @"TICK" () curNES impCurSlotNo
+  impCurSlotNoL += 1
   impNESL .= nes
+
+-- | Win with supplied probability
+drawBoolWithProbability ::
+  HasStatefulGen g m =>
+  -- | Probability with which this action should produce `True`
+  UnitInterval ->
+  m Bool
+drawBoolWithProbability probability = do
+  let p = unboundRational probability
+  n <- uniformRM (1, denominator p)
+  pure (n <= numerator p)
 
 -- | Runs the TICK rule until the next epoch is reached
 passEpoch ::
@@ -1349,21 +1360,29 @@ passEpoch ::
   (ShelleyEraImp era, HasCallStack) =>
   ImpTestM era ()
 passEpoch = do
-  let
-    tickUntilNewEpoch curEpochNo = do
-      oldNES <- getsNES id
-      passTick @era
-      newEpochNo <- getsNES nesELL
-      if newEpochNo > curEpochNo
-        then do
-          globals <- use impGlobalsL
-          newNES <- getsNES id
-          asks itePostEpochBoundaryHook >>= (\f -> f globals (TRC ((), oldNES, newEpochNo)) newNES)
-        else tickUntilNewEpoch curEpochNo
+  globals <- use impGlobalsL
   preNES <- gets impNES
-  let startEpoch = preNES ^. nesELL
-  logDoc $ "Entering " <> ansiExpr (succ startEpoch)
-  tickUntilNewEpoch startEpoch
+  let
+    curEpochNo = preNES ^. nesELL
+    ticksPerSlot =
+      positiveUnitIntervalRelaxToUnitInterval (activeSlotVal (activeSlotCoeff globals))
+    tickUntilNewEpoch = do
+      tickHasBlock <- drawBoolWithProbability ticksPerSlot
+      if tickHasBlock
+        then do
+          oldNES <- getsNES id
+          passTick @era
+          newEpochNo <- getsNES nesELL
+          if newEpochNo > curEpochNo
+            then do
+              newNES <- getsNES id
+              asks itePostEpochBoundaryHook >>= (\f -> f globals (TRC ((), oldNES, newEpochNo)) newNES)
+            else tickUntilNewEpoch
+        else do
+          impCurSlotNoL += 1
+          tickUntilNewEpoch
+  logDoc $ "Entering " <> ansiExpr (succ curEpochNo)
+  tickUntilNewEpoch
   gets impNES >>= epochBoundaryCheck preNES
 
 epochBoundaryCheck ::
@@ -1844,9 +1863,9 @@ produceScript scriptHash = do
 
 advanceToPointOfNoReturn :: ImpTestM era ()
 advanceToPointOfNoReturn = do
-  impLastTick <- gets impLastTick
-  (_, slotOfNoReturn, _) <- runShelleyBase $ getTheSlotOfNoReturn impLastTick
-  impLastTickL .= slotOfNoReturn
+  impCurSlotNo <- gets impCurSlotNo
+  (_, slotOfNoReturn, _) <- runShelleyBase $ getTheSlotOfNoReturn impCurSlotNo
+  impCurSlotNoL .= slotOfNoReturn
 
 -- | A legal ProtVer that differs in the minor Version
 minorFollow :: ProtVer -> ProtVer
@@ -1869,7 +1888,8 @@ whenMajorVersion ::
   , MinVersion <= v
   , v <= MaxVersion
   ) =>
-  ImpTestM era () -> ImpTestM era ()
+  ImpTestM era () ->
+  ImpTestM era ()
 whenMajorVersion a = do
   pv <- getProtVer
   when (pvMajor pv == natVersion @v) a
@@ -1881,7 +1901,8 @@ whenMajorVersionAtLeast ::
   , MinVersion <= v
   , v <= MaxVersion
   ) =>
-  ImpTestM era () -> ImpTestM era ()
+  ImpTestM era () ->
+  ImpTestM era ()
 whenMajorVersionAtLeast a = do
   pv <- getProtVer
   when (pvMajor pv >= natVersion @v) a
@@ -1893,7 +1914,8 @@ whenMajorVersionAtMost ::
   , MinVersion <= v
   , v <= MaxVersion
   ) =>
-  ImpTestM era () -> ImpTestM era ()
+  ImpTestM era () ->
+  ImpTestM era ()
 whenMajorVersionAtMost a = do
   pv <- getProtVer
   when (pvMajor pv <= natVersion @v) a
@@ -1905,7 +1927,8 @@ unlessMajorVersion ::
   , MinVersion <= v
   , v <= MaxVersion
   ) =>
-  ImpTestM era () -> ImpTestM era ()
+  ImpTestM era () ->
+  ImpTestM era ()
 unlessMajorVersion a = do
   pv <- getProtVer
   unless (pvMajor pv == natVersion @v) a